+++ CHANGELOG.md
@@ -1,3 +1,7 @@
 ## 0.1.0 (Unreleased)
 
+BUG FIXES:
+
+* resource/bcm_cmdevice_category: Fixed roles[].uuid computed value never being populated from BCM API response ([#83](https://github.com/hashi-demo-lab/terraform-provider-bcm/issues/83))
+
 FEATURES: